refactor: extract @polpo-ai/file-stores package
Phase 2 of dissolving the root shell into packages. The 15 file-based
store implementations (the default zero-dependency persistence backend)
move from src/stores/ to a proper workspace package, symmetric with
@polpo-ai/drizzle:

- @polpo-ai/file-stores: all File*/Json* stores + the .polpo directory
  layout helpers (POLPO_DIR_NAME, getPolpoDir, getGlobalPolpoDir — the
  layout IS the file persistence contract) + playbook file persistence
  (discovery/CRUD, moved out of src/core/playbook.ts)
- playbook pure logic (validateParams, instantiatePlaybook,
  validatePlaybookDefinition) moves to @polpo-ai/core (playbook-logic),
  where the playbook-store interface already promised it belonged;
  src/core/playbook.ts is gone
- shell consumers (orchestrator, runner, server routes, tests) import
  the package; the polpo-ai public surface re-exports unchanged
- release workflow publishes the new package right after core
