+++ package.json
@@ -31,7 +31,7 @@
     "LICENSE"
   ],
   "scripts": {
-    "build": "pnpm --filter @polpo-ai/core build && pnpm --filter @polpo-ai/vault-crypto build && pnpm --filter @polpo-ai/drizzle build && pnpm --filter @polpo-ai/llm build && pnpm --filter @polpo-ai/tools build && pnpm --filter @polpo-ai/server build && tsc && pnpm --filter @polpo-ai/sdk build && pnpm --filter @polpo-ai/react build && pnpm --filter @polpo-ai/cli build",
+    "build": "pnpm --filter @polpo-ai/core build && pnpm --filter @polpo-ai/file-stores build && pnpm --filter @polpo-ai/vault-crypto build && pnpm --filter @polpo-ai/drizzle build && pnpm --filter @polpo-ai/llm build && pnpm --filter @polpo-ai/tools build && pnpm --filter @polpo-ai/server build && tsc && pnpm --filter @polpo-ai/sdk build && pnpm --filter @polpo-ai/react build && pnpm --filter @polpo-ai/cli build",
     "dev": "tsc --watch",
     "start": "node packages/cli/dist/index.js",
     "test": "vitest",
@@ -68,6 +68,7 @@
     "@hono/node-server": "^1.19.9",
     "@hono/zod-openapi": "^1.2.2",
     "@polpo-ai/core": "workspace:*",
+    "@polpo-ai/file-stores": "workspace:*",
     "@polpo-ai/llm": "workspace:*",
     "@polpo-ai/server": "workspace:*",
     "@polpo-ai/tools": "workspace:*",
