+++ packages/file-stores/src/file-team-store.ts
