+++ src/__tests__/file-team-agent-store.test.ts
@@ -1,8 +1,8 @@
 import { describe, it, expect, beforeEach, afterEach } from "vitest";
 import { rmSync, existsSync, readFileSync } from "node:fs";
 import { join } from "node:path";
-import { FileTeamStore } from "../stores/file-team-store.js";
-import { FileAgentStore } from "../stores/file-agent-store.js";
+import { FileTeamStore } from "@polpo-ai/file-stores";
+import { FileAgentStore } from "@polpo-ai/file-stores";
 import type { Team, AgentConfig } from "@polpo-ai/core/types";
 
 const TEST_DIR = join(process.cwd(), ".test-team-agent-store");
