+++ src/__tests__/json-task-store.test.ts
@@ -1,7 +1,7 @@
 import { describe, it, expect, beforeEach, afterEach } from "vitest";
 import { mkdirSync, rmSync, existsSync, readFileSync } from "node:fs";
 import { join } from "node:path";
-import { JsonTaskStore } from "../stores/json-task-store.js";
+import { JsonTaskStore } from "@polpo-ai/file-stores";
 import type { Task } from "@polpo-ai/core/types";
 
 const TEST_DIR = join(process.cwd(), ".test-orchestra-store");
